v0.4

+++ src/drivers/linux_pwm_out/linux_pwm_out.cpp
@@ -43,7 +43,6 @@
 #include <uORB/topics/actuator_controls.h>
 #include <uORB/topics/actuator_outputs.h>
 #include <uORB/topics/actuator_armed.h>
-#include <uORB/topics/rc_channels.h>
 
 #include <drivers/drv_hrt.h>
 #include <drivers/drv_mixer.h>
@@ -212,7 +211,7 @@ void task_main(int argc, char *argv[])
 		PX4_INFO("Starting PWM output in ocpoc_mmap mode");
 		pwm_out = new OcpocMmapPWMOut(_max_num_outputs);
 
-	} else { /* navio */
+	} else { // navio
 		PX4_INFO("Starting PWM output in Navio mode");
 		pwm_out = new NavioSysfsPWMOut(_device, _max_num_outputs);
 	}
@@ -285,7 +284,7 @@ void task_main(int argc, char *argv[])
 			_controls[0].control[2] = 0.f;
 			int channel = rc_channels.function[rc_channels_s::RC_CHANNELS_FUNCTION_THROTTLE];
 
-			if (ret == 0 && channel >= 0 && channel < (int)(sizeof(rc_channels.channels) / sizeof(rc_channels.channels[0]))) {
+			if (ret == 0 && channel >= 0 && channel < sizeof(rc_channels.channels) / sizeof(rc_channels.channels[0])) {
 				_controls[0].control[3] = rc_channels.channels[channel];
 
 			} else {
@@ -347,7 +346,7 @@ void task_main(int argc, char *argv[])
 					pwm_value = _pwm_min;
 				}
 
-				for (uint32_t i = 0; i < _outputs.noutputs; ++i) {
+				for (int i = 0; i < _outputs.noutputs; ++i) {
 					pwm[i] = pwm_value;
 				}
 
@@ -495,7 +494,7 @@ int linux_pwm_out_main(int argc, char *argv[])
 		}
 	}
 
-	/** gets the parameters for the esc's pwm */
+	// gets the parameters for the esc's pwm
 	param_get(param_find("PWM_DISARMED"), &linux_pwm_out::_pwm_disarmed);
 	param_get(param_find("PWM_MIN"), &linux_pwm_out::_pwm_min);
 	param_get(param_find("PWM_MAX"), &linux_pwm_out::_pwm_max);
